moved loading view into core library
