+++ storybook/helpers/ComponentView.tsx
@@ -9,7 +9,6 @@ import {
   Divider
 } from '@material-ui/core';
 
-import LoadingView from './LoadingView';
 import { ComponentLogs } from "../../src";
 import { PropertyEditors, PropertyData, PropertyType } from "./PropertyEditors";
 export { PropertyData, PropertyType };
@@ -56,29 +55,26 @@ export default class ComponentView extends React.Component<Props, State> {
           <Component.Entity>
           <Component.Data>
           <Component.Logs>
-          {(entity: any, data: any, logs: ComponentLogs) => {
-            if ( !entity || !data)
-              return <LoadingView logs={logs}/>
-            return (
-              <>
-              <Typography variant="h6" component="h6">
-                Entity
-              </Typography>
-              {objectInspector(entity, `${name}.Entity`, "Entity Instance")}
-              <Typography variant="h6" component="h6">
-                Data
-              </Typography>
-              {objectInspector(data, `${name}.Data`, "Semantic Data")}
-              <Typography variant="h6" component="h6">
-                Logs
-              </Typography>
-              {objectInspector(logs.react, `${name}.ReactLogs`, "React Logs")}
-              {objectInspector(logs.entity, `${name}.EntityLogs`, "Entity Logs")}
-              {objectInspector(logs.data, `${name}.DataLogs`, "Data Query Logs")}
-              {objectInspector(logs.code, `${name}.CodeLogs`, "Code Logs")}
-              {objectInspector(logs.prose, `${name}.ProseLogs`, "Prose Logs")}
-              </>
-            )}}
+          {(entity: any, data: any, logs: ComponentLogs) => (
+            <>
+            <Typography variant="h6" component="h6">
+              Entity
+            </Typography>
+            {objectInspector(entity, `${name}.Entity`, "Entity Instance")}
+            <Typography variant="h6" component="h6">
+              Data
+            </Typography>
+            {objectInspector(data, `${name}.Data`, "Semantic Data")}
+            <Typography variant="h6" component="h6">
+              Logs
+            </Typography>
+            {objectInspector(logs.react, `${name}.ReactLogs`, "React Logs")}
+            {objectInspector(logs.entity, `${name}.EntityLogs`, "Entity Logs")}
+            {objectInspector(logs.data, `${name}.DataLogs`, "Data Query Logs")}
+            {objectInspector(logs.code, `${name}.CodeLogs`, "Code Logs")}
+            {objectInspector(logs.prose, `${name}.ProseLogs`, "Prose Logs")}
+            </>
+          )}
           </Component.Logs>
           </Component.Data>
           </Component.Entity>
